aider: Added a login route to handle form submission, checking if the submitted username and password match a record in the users table.
# Context:
USER: shoudl we add the login route app.py as well?
ASSISTANT: Yes, we should add a route to handle the login form submission. This route will check if the submitted username and password match a record in the `users` table in the `aitools_ebookwriter` database.

```python
app.py
<<<<<<< ORIGINAL
@app.route('/')
def login():
    return render_template('login.html')

@app.route('/home')
def home():
    return render_template('home.html')

if __name__ == '__main__':
    app.run(debug=True)
=======
@app.route('/', methods=['GET', 'POST'])
def login():
    if request.method == 'POST':
        username = request.form.get('username')
        password = request.form.get('password')

        cursor = mysql.connection.cursor()
        cursor.execute(''' SELECT * FROM users WHERE username = %s AND password = %s ''', (username, password))
        user = cursor.fetchone()
        cursor.close()

        if user:
            return redirect(url_for('home'))
        else:
            return 'Invalid username or password'
    else:
        return render_template('login.html')

@app.route('/home')
def home():
    return render_template('home.html')

if __name__ == '__main__':
    app.run(debug=True)
>>>>>>> UPDATED
```

This new route, `login`, will be triggered when the form in the `login.html` template is submitted. It will extract the form data, query the `users` table with this data, and if a matching user is found, redirect the user to the home page. If no matching user is found, it will return an error message.