aider: Add a link to the bottom of the login.html page to go to the register.html page.
# Context:
USER: Add a link to the bottom of the login.html page to got to the register.html page
ASSISTANT: Sure, I can help with that. Here is the HTML code to add a link to the `register.html` page at the bottom of the `login.html` page.

```html
templates/login.html
<<<<<<< ORIGINAL
<!DOCTYPE html>
<html>
<head>
    <title>Login</title>
</head>
<body>
    <h1>Login</h1>
    <form action="/" method="post">
        <label for="username">Username:</label><br>
        <input type="text" id="username" name="username" required><br>
        <label for="password">Password:</label><br>
        <input type="password" id="password" name="password" required><br>
        <input type="submit" value="Login">
    </form>
</body>
</html>
=======
<!DOCTYPE html>
<html>
<head>
    <title>Login</title>
</head>
<body>
    <h1>Login</h1>
    <form action="/" method="post">
        <label for="username">Username:</label><br>
        <input type="text" id="username" name="username" required><br>
        <label for="password">Password:</label><br>
        <input type="password" id="password" name="password" required><br>
        <input type="submit" value="Login">
    </form>
    <p>Don't have an account? <a href="/register">Register</a></p>
</body>
</html>
>>>>>>> UPDATED
```

This change adds a paragraph at the bottom of the `login.html` page with a link to the `register.html` page. The text of the link is "Register".