aider: Add include header.html to home.html.
# Context:
USER: Check first and if missing, Add a include header.html to each web pages that contains the header and associated CSS to create a Dark Mode theme. The title of the application is Ebook Wizard. The header should also contains a button that will logout and deauthroise the current user
ASSISTANT: To add the header to each web page, we need to modify the following files:

- templates\home.html
- templates\login.html
- templates\register.html

However, these files are currently read-only. Could you please make them read-write so that I can propose the necessary changes?
USER: check home.html for included header.html
ASSISTANT: The `home.html` file does not currently include `header.html`. Here's how we can include it:

```html
templates\home.html
<<<<<<< ORIGINAL
<!DOCTYPE html>
<html>
<head>
    <title>Home</title>
</head>
<body>
    <h1>Welcome to the Home Page!</h1>
</body>
</html>
=======
<!DOCTYPE html>
<html>
<head>
    <title>Home</title>
</head>
<body>
    {% include 'header.html' %}
    <h1>Welcome to the Home Page!</h1>
</body>
</html>
>>>>>>> UPDATED
```

This change uses the `{% include 'header.html' %}` directive to include the contents of `header.html` in `home.html`. This will add the header and associated CSS to the home page.