aider: Import 'cursors' module from 'MySQLdb' and use it to create a dictionary cursor.
# Context:
USER: AttributeError
AttributeError: type object 'MySQL' has no attribute 'cursors'
ASSISTANT: The error message indicates that the MySQL object does not have an attribute named 'cursors'. This is likely because the 'cursors' attribute is not part of the 'MySQL' class in the 'flask_mysqldb' module.

To fix this, we need to import the 'cursors' module from 'MySQLdb' and use it to create a dictionary cursor. Here is how to do it:

```python
app.py
<<<<<<< ORIGINAL
    cursor = mysql.connection.cursor(MySQL.cursors.DictCursor)
=======
    from MySQLdb import cursors
    cursor = mysql.connection.cursor(cursors.DictCursor)
>>>>>>> UPDATED
```

This change should fix the AttributeError.