aider: Fix project data display on projects.html by accessing the project data using dictionary keys instead of indices.
# Context:
USER: no project data is displayed on projects.html but the data is being retrieved from the databse
ASSISTANT: The issue might be related to how the project data is being accessed in the `projects.html` file. The project data is fetched as a list of dictionaries, where each dictionary represents a project. In the `projects.html` file, you are trying to access the project data using indices, which is incorrect.

Here's how to fix it:

```html
templates\projects.html
<<<<<<< ORIGINAL
            <td style="padding: 15px; text-align: left;">{{ project[1] }}</td>
            <td style="padding: 15px; text-align: left;">{{ project[2] }}</td>
            <td style="padding: 15px; text-align: left;">{{ project[3] }}</td>
            <td style="padding: 15px; text-align: left;">{{ project[4] }}</td>
            <td style="padding: 15px; text-align: left;">{{ project[6] }}</td>
=======
            <td style="padding: 15px; text-align: left;">{{ project.Title }}</td>
            <td style="padding: 15px; text-align: left;">{{ project.Description }}</td>
            <td style="padding: 15px; text-align: left;">{{ project.AuthorRole }}</td>
            <td style="padding: 15px; text-align: left;">{{ project.AuthorTone }}</td>
            <td style="padding: 15px; text-align: left;">{{ project.InitialPrompt }}</td>
>>>>>>> UPDATED
```

This change should fix the issue. The project data should now be displayed correctly on the `projects.html` page.