aider: Add button to the end of each project record in projects.html to route it to edit_project.html for editing/updating the project details.
# Context:
USER: add button to the end of each project record that is displayed in projects.html - route it to edit_project.html that allows editing/updating of the project details
ASSISTANT: To add an "Edit" button to each project record, we need to modify the `templates\projects.html` file.

Here's the change:

```html
templates\projects.html
<<<<<<< ORIGINAL
            <td style="padding: 15px; text-align: left;">{{ project.InitialPrompt }}</td>
        </tr>
        {% endfor %}
    </table>
=======
            <td style="padding: 15px; text-align: left;">{{ project.InitialPrompt }}</td>
            <td style="padding: 15px; text-align: left;">
                <button onclick="location.href='edit_project/{{ project.id }}'">Edit</button>
            </td>
        </tr>
        {% endfor %}
    </table>
>>>>>>> UPDATED
```

This change adds a new table cell to each row in the projects table. Inside this cell, there's a button that, when clicked, will redirect the user to the `edit_project` route for the corresponding project. The project's id is included in the URL to identify which project to edit.

Please note that this change assumes that you have an `edit_project` route set up in your Flask application that accepts a project id as a parameter, and that each project has an `id` attribute. If this is not the case, please let me know so we can adjust the solution accordingly.