aider: Created edit_project.html template for editing project details.
# Context:
USER: 127.0.0.1 - - [16/Jul/2023 10:20:35] "GET /edit_project/ HTTP/1.1" 404 -
ASSISTANT: The 404 error indicates that the server couldn't find the requested URL. In this case, it seems like you're trying to access the `/edit_project/` route without providing a project id.

The `edit_project` route is defined as `/edit_project/<int:project_id>`, which means it expects an integer project id as part of the URL. If you try to access `/edit_project/` without a project id, you'll get a 404 error because that route doesn't exist.

To fix this, you should make sure to include a project id when you try to access the `edit_project` route. For example, if you want to edit the project with id 1, you would access the URL `/edit_project/1`.
USER: TemplateNotFound
jinja2.exceptions.TemplateNotFound: edit_project.html
ASSISTANT: The error message indicates that Flask is unable to find the `edit_project.html` template. This could be because the file doesn't exist, or because it's not in the correct location.

Flask looks for templates in a folder named `templates` in the same directory as the script that's running. If your `edit_project.html` file is not in this folder, Flask won't be able to find it.

To fix this issue, you need to create the `edit_project.html` file in the `templates` directory. If you already have this file but it's in a different location, you should move it to the `templates` directory.

Here's an example of what the `edit_project.html` file could look like:

```html
templates\edit_project.html
<<<<<<< ORIGINAL
=======
<!DOCTYPE html>
<html>
<head>
    <title>Edit Project</title>
</head>
<body>
    <h1>Edit Project</h1>
    <!-- Add your form here to edit the project details -->
</body>
</html>
>>>>>>> UPDATED
```

This is a very basic template that just displays a title. You'll need to add a form to this template to allow the user to edit the project details. The form should include fields for all the project attributes that the user should be able to edit, and it should submit a POST request to the `edit_project` route when the user submits the form.