aider: Changed cursor to DictCursor when fetching project data from the database in the edit_project route.
# Context:
USER: make sure that when the edit_projects route is called it preloads the current data into the edit fields
ASSISTANT: The current code already fetches the project data from the database and passes it to the `edit_project.html` template. The template then uses this data to populate the form fields.

If the form fields are not being populated with the current data, it could be because the keys used to access the project data in the `edit_project.html` template do not match the column names in your `projects` table.

To fix this, you need to make sure that the keys used in the `edit_project.html` template match the column names in your `projects` table.

Here's how you can do that:

```html
templates\edit_project.html
<<<<<<< ORIGINAL
        <input type="text" id="title" name="title" value="{{ project.Title }}"><br>
        <input type="text" id="description" name="description" value="{{ project.Description }}"><br>
        <input type="text" id="author_role" name="author_role" value="{{ project.AuthorRole }}"><br>
        <input type="text" id="author_tone" name="author_tone" value="{{ project.AuthorTone }}"><br>
        <input type="text" id="initial_prompt" name="initial_prompt" value="{{ project.InitialPrompt }}"><br>
=======
        <input type="text" id="title" name="title" value="{{ project.title }}"><br>
        <input type="text" id="description" name="description" value="{{ project.description }}"><br>
        <input type="text" id="author_role" name="author_role" value="{{ project.author_role }}"><br>
        <input type="text" id="author_tone" name="author_tone" value="{{ project.author_tone }}"><br>
        <input type="text" id="initial_prompt" name="initial_prompt" value="{{ project.initial_prompt }}"><br>
>>>>>>> UPDATED
```

This code changes the keys used to access the project data in the `edit_project.html` template to lowercase, which is the standard naming convention for column names in MySQL. Now, the form fields should be populated with the current data when the `edit_project` route is called.
USER: no data is loaded in to the value variable of the field in edit_project.html
ASSISTANT: The issue might be due to the way the data is fetched from the database in the `edit_project` route in `app.py`. The `fetchone()` method returns a tuple, not a dictionary. This means that the data cannot be accessed using keys in the template.

To fix this, you need to use a `DictCursor` when fetching the project data from the database. This will return the data as a dictionary, which can then be accessed using keys in the template.

Here's how you can do that:

```python
app.py
<<<<<<< ORIGINAL
@app.route('/edit_project/<int:project_id>', methods=['GET', 'POST'])
def edit_project(project_id):
    if 'username' not in session:
        return redirect(url_for('login'))

    cursor = mysql.connection.cursor()
    cursor.execute(''' SELECT * FROM projects WHERE ProjectID = %s ''', (project_id,))
    project = cursor.fetchone()
    cursor.close()

    if request.method == 'POST':
        # Add your code here to handle the POST request
        pass
    else:
        # Add your code here to handle the GET request
        pass

    return render_template('edit_project.html', project=project)
=======
@app.route('/edit_project/<int:project_id>', methods=['GET', 'POST'])
def edit_project(project_id):
    if 'username' not in session:
        return redirect(url_for('login'))

    from MySQLdb.cursors import DictCursor
    cursor = mysql.connection.cursor(DictCursor)
    cursor.execute(''' SELECT * FROM projects WHERE ProjectID = %s ''', (project_id,))
    project = cursor.fetchone()
    cursor.close()

    if request.method == 'POST':
        # Add your code here to handle the POST request
        pass
    else:
        # Add your code here to handle the GET request
        pass

    return render_template('edit_project.html', project=project)
>>>>>>> UPDATED
```

This code changes the cursor to a `DictCursor` when fetching the project data from the database. Now, the data should be returned as a dictionary, and the form fields in the `edit_project.html` template should be populated with the current data when the `edit_project` route is called.