aider: Add header and CSS formatting to edit_project.html.
# Context:
USER: add header and css formatting to edit_project.html
ASSISTANT: To add a header and CSS formatting to the `edit_project.html` file, you need to include a link to a CSS file in the `<head>` section of the HTML and include the header file using Jinja2's `include` statement.

Here's how you can do that:

```html
templates\edit_project.html
<<<<<<< ORIGINAL
<head>
    <title>Edit Project</title>
</head>
<body>
    <h1>Edit Project</h1>
=======
<head>
    <title>Edit Project</title>
    <link rel="stylesheet" type="text/css" href="{{ url_for('static', filename='styles.css') }}">
</head>
<body>
    {% include 'header.html' %}
    <h1>Edit Project</h1>
>>>>>>> UPDATED
```

This code includes a link to a CSS file named `styles.css` in the `static` directory and includes the `header.html` file at the top of the body.

Please note that you need to create the `styles.css` file in the `static` directory and add your CSS rules there. If the `static` directory or the `styles.css` file does not exist, you need to create them.

If you need help with creating the `styles.css` file or adding CSS rules, please let me know.